# userService/src/main/java/com/neyma/userService/dto/UserResponse.java
package com.neyma.userService.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import java.time.LocalDateTime;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UserResponse {
    private java.util.UUID id;
    private String name;
    private String email;
    private LocalDateTime lastLoginTime;
}