# userService/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: User Service API
  description: API for managing users, including registration and login.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
paths:
  /users/register:
    post:
      summary: Register a new user
      description: Creates a new user account with the provided name, email, and password.
      operationId: register
      requestBody:
        description: User registration details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                example:
                  email: "must be a well-formed email address"
                  password: "Password is required"
        '409':
          description: User with the given email already exists
          content:
            text/plain:
              schema:
                type: string
                example: User with email example@test.com already exists
        '415':
          description: Unsupported Media Type
          content:
            text/plain:
              schema:
                type: string
                example: "Content-Type 'text/plain;charset=UTF-8' is not supported. Please use 'Content-Type: application/json'"
  /users/login:
    post:
      summary: Login user
      description: Authenticates a user with email and password.
      operationId: login
      requestBody:
        description: User login credentials
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: User successfully logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                example:
                  email: "must be a well-formed email address"
        '401':
          description: Invalid email or password
          content:
            text/plain:
              schema:
                type: string
                example: Invalid email or password
        '415':
          description: Unsupported Media Type
          content:
            text/plain:
              schema:
                type: string
                example: "Content-Type 'text/plain;charset=UTF-8' is not supported. Please use 'Content-Type: application/json'"
  /users/search:
    get:
      summary: Search users
      description: Search for users by partial name or email (case-insensitive).
      operationId: searchUsers
      parameters:
        - name: query
          in: query
          description: Search query string for name or email
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of users matching the query
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '400':
          description: Missing query parameter
  /users/{id}:
    get:
      summary: Get user by ID
      description: Retrieves user details by their ID.
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            text/plain:
              schema:
                type: string
                example: User not found
components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: securepassword123
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: securepassword123
      required:
        - email
        - password
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        lastLoginTime:
          type: string
          format: date-time
          example: '2023-10-27T10:00:00Z'